refactor: ユーザー情報取得をFirebaseからstoreに切り替え
